pyproject.toml
[tool.poetry]
name = "recommender"
version = "0.1.0"
description = ""
authors = ["tobiadefami <tobiadefami@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.10.0"
asyncio = "^3.4.3"
fastapi = "^0.111.1"
sqlalchemy = { extras = ["asyncio"], version = "^2.0.36" }
psycopg2-binary = "^2.9.1"
youtube-transcript-api = "^0.6.2"
google-api-python-client = "^2.145.0"
asyncpraw = "^7.7.1"
fastapi-cors = "^0.0.6"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = "^1.7.4"
langchain = "^0.3.9"
langchain-core = "^0.3.21"
langchain-openai = "^0.2.10"
langchain-community = "^0.3.8"
langchain-google-community = "^2.0.3"
duckduckgo-search = "^6.3.7"
exa-py = "^1.7.0"
langchain-exa = "^0.2.1"
asyncpg = "^0.30.0"
ruff = "^0.9.1"
langchain-google-vertexai = "^2.0.12"

[tool.pyright]
venvPath = "."
venv = "/home/chief/.pyenv/versions/3.11.11/envs/recommender_dev"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
